remove bunch of unnesscary async/awiats. Seperate concerns
