+++ samples/bot-calling-meeting/csharp/Source/CallingMediaBot.Web/CallingMediaBot.Web.csproj
@@ -4,6 +4,8 @@
     <TargetFramework>net6.0</TargetFramework>
     <ImplicitUsings>enable</ImplicitUsings>
     <Nullable>enable</Nullable>
+    <EnforceCodeStyleInBuild>true</EnforceCodeStyleInBuild>
+    <AnalysisLevel>preview-recommended</AnalysisLevel>
   </PropertyGroup>
 
   <ItemGroup>
@@ -12,7 +14,6 @@
     <PackageReference Include="Azure.Identity" Version="1.7.0" />
     <PackageReference Include="Microsoft.Graph.Communications.Core" Version="1.2.0.5304" />
     <PackageReference Include="Microsoft.Graph.Communications.Calls" Version="1.2.0.5304" />
-    <PackageReference Include="System.Net.Http.Formatting.Extension" Version="5.2.3" />
     <PackageReference Include="AdaptiveCards" Version="2.7.3" />
     <PackageReference Include="AdaptiveCards.Templating" Version="1.3.1" />
   </ItemGroup>
