+++ samples/bot-calling-meeting/csharp/Source/CallingMediaBot.Web/Controllers/CallingCallbackController.cs
@@ -1,0 +1,21 @@
+﻿using CallingMediaBot.Web.Bots;
+using Microsoft.AspNetCore.Mvc;
+
+namespace CallingMeetingBot.Controllers;
+
+[Route("callback")]
+public class CallingCallbackController : Controller
+{
+    private readonly CallingBot bot;
+
+    public CallingCallbackController(CallingBot bot)
+    {
+        this.bot = bot;
+    }
+
+    [HttpPost, HttpGet]
+    public async Task HandleCallbackRequestAsync()
+    {
+        await bot.ProcessNotificationAsync(Request, Response).ConfigureAwait(false);
+    }
+}
