+++ samples/bot-calling-meeting/csharp/Source/CallingMediaBot.Web/Utility/CommonUtils.cs
@@ -12,9 +12,9 @@ namespace CallingMediaBot.Web.Utility
         public static async Task ForgetAndLogExceptionAsync(
             this Task task,
             IGraphLogger logger,
-            string description = null,
-            [CallerMemberName] string memberName = null,
-            [CallerFilePath] string filePath = null,
+            string? description = null,
+            [CallerMemberName] string? memberName = null,
+            [CallerFilePath] string? filePath = null,
             [CallerLineNumber] int lineNumber = 0)
         {
             try
